update bin_to_ascii to use doubles, add gen_vecs to create input data for dataflow/spark microbenchmarks (see dataflow.git)

+++ binary_convert/bin_to_ascii.cpp
@@ -2,8 +2,8 @@
 
 int main(int argc,char**argv)
 {
-  float vec[3];
-  while (std::cin.read(reinterpret_cast<char*>(vec),3*sizeof(float)))
+  double vec[3];
+  while (std::cin.read(reinterpret_cast<char*>(vec),3*sizeof(double)))
     std::cout<<vec[0]<<" "<<vec[1]<<" "<<vec[2]<<std::endl;
 
   return 0;
